+++ bindings/python/crocoddyl/core/actuation-base.hpp
@@ -23,15 +23,15 @@ class ActuationModelAbstract_wrap : public ActuationModelAbstract, public bp::wr
 
   void calc(const boost::shared_ptr<ActuationDataAbstract>& data, const Eigen::Ref<const Eigen::VectorXd>& x,
             const Eigen::Ref<const Eigen::VectorXd>& u) {
-    assert(x.size() == state_.get_nx() && "x has wrong dimension");
-    assert(u.size() == nu_ && "u has wrong dimension");
+    assert(static_cast<std::size_t>(x.size()) == state_.get_nx() && "x has wrong dimension");
+    assert(static_cast<std::size_t>(u.size()) == nu_ && "u has wrong dimension");
     return bp::call<void>(this->get_override("calc").ptr(), data, (Eigen::VectorXd)x, (Eigen::VectorXd)u);
   }
 
   void calcDiff(const boost::shared_ptr<ActuationDataAbstract>& data, const Eigen::Ref<const Eigen::VectorXd>& x,
                 const Eigen::Ref<const Eigen::VectorXd>& u, const bool& recalc = true) {
-    assert(x.size() == state_.get_nx() && "x has wrong dimension");
-    assert(u.size() == nu_ && "u has wrong dimension");
+    assert(static_cast<std::size_t>(x.size()) == state_.get_nx() && "x has wrong dimension");
+    assert(static_cast<std::size_t>(u.size()) == nu_ && "u has wrong dimension");
     return bp::call<void>(this->get_override("calcDiff").ptr(), data, (Eigen::VectorXd)x, (Eigen::VectorXd)u, recalc);
   }
 };
