+++ src/core/activations/smooth-abs.cpp
@@ -16,7 +16,7 @@ ActivationModelSmoothAbs::~ActivationModelSmoothAbs() {}
 
 void ActivationModelSmoothAbs::calc(const boost::shared_ptr<ActivationDataAbstract>& data,
                                     const Eigen::Ref<const Eigen::VectorXd>& r) {
-  assert(r.size() == nr_ && "r has wrong dimension");
+  assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
   boost::shared_ptr<ActivationDataSmoothAbs> d = boost::static_pointer_cast<ActivationDataSmoothAbs>(data);
 
   d->a = (r.array().cwiseAbs2().array() + 1).array().cwiseSqrt();
@@ -25,7 +25,7 @@ void ActivationModelSmoothAbs::calc(const boost::shared_ptr<ActivationDataAbstra
 
 void ActivationModelSmoothAbs::calcDiff(const boost::shared_ptr<ActivationDataAbstract>& data,
                                         const Eigen::Ref<const Eigen::VectorXd>& r, const bool& recalc) {
-  assert(r.size() == nr_ && "r has wrong dimension");
+  assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
   if (recalc) {
     calc(data, r);
   }
