+++ src/multibody/contact-base.cpp
@@ -20,8 +20,11 @@ ContactModelAbstract::~ContactModelAbstract() {}
 
 void ContactModelAbstract::updateForceDiff(const boost::shared_ptr<ContactDataAbstract>& data,
                                            const Eigen::MatrixXd& df_dx, const Eigen::MatrixXd& df_du) const {
-  assert((df_dx.rows() == nc_ || df_dx.cols() == state_.get_nx()) && "df_dx has wrong dimension");
-  assert((df_du.rows() == nc_ || df_du.cols() == nu_) && "df_du has wrong dimension");
+  assert(
+      (static_cast<std::size_t>(df_dx.rows()) == nc_ || static_cast<std::size_t>(df_dx.cols()) == state_.get_nx()) &&
+      "df_dx has wrong dimension");
+  assert((static_cast<std::size_t>(df_du.rows()) == nc_ || static_cast<std::size_t>(df_du.cols()) == nu_) &&
+         "df_du has wrong dimension");
   data->df_dx = df_dx;
   data->df_du = df_du;
 }
